+++ src/ble/reducers.test.ts
@@ -3,14 +3,6 @@
 
 import { AnyAction } from 'redux';
 import {
-    bleDIServiceDidReceiveFirmwareRevision,
-    bleDIServiceDidReceivePnPId,
-} from '../ble-device-info-service/actions';
-import { PnpIdVendorIdSource } from '../ble-device-info-service/protocol';
-import { HubType, LegoCompanyId } from '../ble-lwp3-service/protocol';
-import { didReceiveStatusReport } from '../ble-pybricks-service/actions';
-import { Status, statusToFlag } from '../ble-pybricks-service/protocol';
-import {
     bleConnectPybricks,
     bleDidConnectPybricks,
     bleDidDisconnectPybricks,
@@ -26,11 +18,6 @@ test('initial state', () => {
     expect(reducers(undefined, {} as AnyAction)).toMatchInlineSnapshot(`
         {
           "connection": "ble.connection.state.disconnected",
-          "deviceBatteryCharging": false,
-          "deviceFirmwareVersion": "",
-          "deviceLowBatteryWarning": false,
-          "deviceName": "",
-          "deviceType": "",
         }
     `);
 });
@@ -72,88 +59,4 @@ test('connection', () => {
             bleDidFailToDisconnectPybricks(),
         ).connection,
     ).toBe(BleConnectionState.Connected);
-});
-
-test('deviceName', () => {
-    const testId = 'test-id';
-    const testName = 'Test Name';
-
-    expect(
-        reducers({ deviceName: '' } as State, bleDidConnectPybricks(testId, testName))
-            .deviceName,
-    ).toBe(testName);
-
-    expect(
-        reducers({ deviceName: testName } as State, bleDidDisconnectPybricks())
-            .deviceName,
-    ).toBe('');
-});
-
-test('deviceType', () => {
-    expect(
-        reducers(
-            { deviceType: '' } as State,
-            bleDIServiceDidReceivePnPId({
-                vendorIdSource: PnpIdVendorIdSource.BluetoothSig,
-                vendorId: LegoCompanyId,
-                productId: HubType.MoveHub,
-                productVersion: 0,
-            }),
-        ).deviceType,
-    ).toBe('Move hub');
-
-    expect(
-        reducers({ deviceType: 'Move hub' } as State, bleDidDisconnectPybricks())
-            .deviceType,
-    ).toBe('');
-});
-
-test('deviceFirmwareVersion', () => {
-    const testVersion = '3.0.0';
-
-    expect(
-        reducers(
-            { deviceFirmwareVersion: '' } as State,
-            bleDIServiceDidReceiveFirmwareRevision(testVersion),
-        ).deviceFirmwareVersion,
-    ).toBe(testVersion);
-
-    expect(
-        reducers(
-            { deviceFirmwareVersion: testVersion } as State,
-            bleDidDisconnectPybricks(),
-        ).deviceFirmwareVersion,
-    ).toBe('');
-});
-
-test('deviceLowBatteryWarning', () => {
-    expect(
-        reducers(
-            { deviceLowBatteryWarning: false } as State,
-            didReceiveStatusReport(statusToFlag(Status.BatteryLowVoltageWarning), 0, 0),
-        ).deviceLowBatteryWarning,
-    ).toBeTruthy();
-
-    expect(
-        reducers(
-            { deviceLowBatteryWarning: true } as State,
-            didReceiveStatusReport(
-                ~statusToFlag(Status.BatteryLowVoltageWarning),
-                0,
-                0,
-            ),
-        ).deviceLowBatteryWarning,
-    ).toBeFalsy();
-
-    expect(
-        reducers({ deviceLowBatteryWarning: true } as State, bleDidDisconnectPybricks())
-            .deviceLowBatteryWarning,
-    ).toBeFalsy();
-});
-
-test('deviceBatteryCharging', () => {
-    expect(
-        reducers({ deviceBatteryCharging: true } as State, bleDidDisconnectPybricks())
-            .deviceBatteryCharging,
-    ).toBeFalsy();
 });
