+++ src/status-bar/StatusBar.test.tsx
@@ -4,7 +4,6 @@
 import { act, waitFor } from '@testing-library/react';
 import React from 'react';
 import { testRender } from '../../test';
-import { BleConnectionState } from '../ble/reducers';
 import { HubRuntimeState } from '../hub/reducers';
 import StatusBar from './StatusBar';
 
@@ -12,15 +11,14 @@ it('should show popover when hub name is clicked', async () => {
     const testHubName = 'Test hub';
 
     const [user, statusBar] = testRender(<StatusBar />, {
-        ble: {
-            connection: BleConnectionState.Connected,
+        hub: {
+            runtime: HubRuntimeState.Idle,
             deviceName: testHubName,
             deviceType: 'hub type',
             deviceFirmwareVersion: 'v0.0.0',
             deviceLowBatteryWarning: false,
             deviceBatteryCharging: false,
         },
-        hub: { runtime: HubRuntimeState.Idle },
     });
 
     await act(() => user.click(statusBar.getByText(testHubName)));
@@ -32,15 +30,14 @@ it('should show popover when battery is clicked', async () => {
     const testHubName = 'Test hub';
 
     const [user, statusBar] = testRender(<StatusBar />, {
-        ble: {
-            connection: BleConnectionState.Connected,
+        hub: {
+            runtime: HubRuntimeState.Idle,
             deviceName: testHubName,
             deviceType: 'hub type',
             deviceFirmwareVersion: 'v0.0.0',
             deviceLowBatteryWarning: false,
             deviceBatteryCharging: false,
         },
-        hub: { runtime: HubRuntimeState.Idle },
     });
 
     await act(() => user.click(statusBar.getByTitle('Battery')));
